Création dans la classe Professeur de la variable boolean activeTrace pour activer ou non le traçage du programme de la variable boolean irc pour switcher mode console au mode IRC

+++ src/edu/projet/expressions/Constante.java
@@ -42,7 +42,7 @@ public class Constante extends Expression {
 		// pour supprimer le point si entier à afficher
 		double d = Double.parseDouble(this.constante);
 		
-		if ( (double) (int) d == d ) {
+		if ( (int) d == d ) {
 			
 			if (d < 0)
 				return "(" + Integer.toString((int) d ) + ")";
